requirements.txt
python-telegram-bot==13.11
pymupdf==1.21.1
fuzzywuzzy==0.18.0
python-Levenshtein==0.12.2  # fuzzywuzzy dependency for improved performance
